+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+## [Unreleased]
+
+### Changed
+
+- Support new runtime protocol session configuration
+
 ## [1.13.1] - 2024-06-05
 
 ### Fixed